scraper/requirements.txt
yt-dlp==2024.12.3
supabase==2.9.1
python-dotenv==1.0.1
requests==2.31.0
